Compare release targets against last release tag, not last commit
Release workflows now compare Bazel targets against the last release
of that specific package (via git tags) rather than HEAD~1. This avoids
unnecessary releases when unrelated files change.

Changes:
- Add unified tools/ci/bazel_diff.py supporting CI and release modes
- Add .github/actions/check-release-needed for release workflows
- Update release workflows with check job that gates release creation
- Remove path-based triggers from release workflows (use bazel-diff)
- Add force_release workflow_dispatch input for manual overrides
