+++ .github/workflows/gnome-terminal-profile-switcher-release.yml
@@ -1,12 +1,18 @@
 # Build and release gnome-terminal-profile-switcher wheel
+#
+# Only creates a release if there are actual Bazel target changes since the last release.
 name: gnome-terminal-profile-switcher Release
 
 on:
   push:
     branches: [devel, main]
-    paths:
-      - "gnome_terminal_profile_switcher/**"
   workflow_dispatch:
+    inputs:
+      force_release:
+        description: "Force release even if no changes detected"
+        required: false
+        default: false
+        type: boolean
 
 # Cancel outdated workflow runs when new commits are pushed
 concurrency:
@@ -17,7 +23,42 @@ permissions:
   contents: write
 
 jobs:
+  check:
+    name: Check if release needed
+    runs-on: ubuntu-latest
+    outputs:
+      release_needed: ${{ steps.check.outputs.release_needed }}
+      reason: ${{ steps.check.outputs.reason }}
+    steps:
+      - name: Check out code
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+
+      - uses: ./.github/actions/bazel-cache
+        id: bazel-cache
+
+      - name: Check if release needed
+        id: check
+        uses: ./.github/actions/check-release-needed
+        with:
+          package_prefix: gnome-terminal-profile-switcher
+          bazel_target_pattern: "//gnome_terminal_profile_switcher/..."
+
+      - name: Log result
+        run: |
+          echo "Release needed: ${{ steps.check.outputs.release_needed }}"
+          echo "Reason: ${{ steps.check.outputs.reason }}"
+
+      - uses: ./.github/actions/bazel-cache-save
+        if: always()
+        with:
+          cache-hit: ${{ steps.bazel-cache.outputs.cache-hit }}
+          cache-key: ${{ steps.bazel-cache.outputs.cache-key }}
+
   release:
+    needs: [check]
+    if: needs.check.outputs.release_needed == 'true' || inputs.force_release == true
     uses: ./.github/workflows/python-wheel-release.yml
     with:
       package_name: gnome-terminal-profile-switcher
